+++ docs/CHANGELOG.md
@@ -21,6 +21,11 @@ Changes are grouped as follows:
   were caused by. In addition, the `NeatSession` object now ensures a linear provenance path for the
   data modeling.
 
+### Added
+- Writing a model to Excel now includes the reference data model if the model was created from another model
+  using the `prepare.data_model.to_enterprise(...)`, `neat.prepare.data_model.to_solution(...)`, or 
+  `neat.prepare.data_model.to_data_product(...)` methods.
+
 ## [0.103.1] - 17-12-**2024**
 ### Fixed
 - Writing a model to Excel or YAML no longer skips specifying version and space of Views unless they both match.